Named the interface back to its original naming as to not affect code earlier/later in the process that we are unaware of.
Worked out how to split the difference hour intervals into periods, and calculate them separately based on that.

Removed TollFeeResult and TollPass and Bill because this was not requested and i realized that it was probably outside of the bounds of this assignment.

I Consider my version of the GetTollFee to be easier to understand and read then the initial version i was supplied, also far superior to my earlier iterations.